requirements.txt
slack-bolt==1.24.0
slack-sdk==3.36.0
pillow==11.3.0
requests==2.32.5
bedrock-agentcore==0.1.2
mcp==1.13.1
pytz==2025.2
strands-agents==1.7.1
strands-agents-tools==0.2.6

# avoid fastuuid dependency
# see: https://github.com/BerriAI/litellm/issues/14145
litellm==1.76.0

# workaround for https://github.com/BerriAI/litellm/issues/13081
apscheduler
backoff
cryptography
email-validator
fastapi
fastapi-sso
orjson
websockets
